# Improve template rendering performance on Rust