--- a/frameworks/Rust/actix/Cargo.toml
+++ b/frameworks/Rust/actix/Cargo.toml
@@ -33,7 +33,6 @@ actix-server = "1.0.4"
 actix-service = "1.0.6"
 snmalloc-rs = "0.2.6"
 askama = "0.8"
-yarte = { version = "0.12", features = ["bytes-buf"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 env_logger = "0.7"
@@ -50,6 +49,7 @@ log = { version = "0.4", features = ["release_max_level_debug"] }
 v_htmlescape = "0.10"
 tokio = "0.2"
 tokio-postgres = { git = "https://github.com/fafhrd91/rust-postgres.git" }
+sailfish = "0.3.0"
 
 [build-dependencies]
 askama = "0.8"