--- a/frameworks/Rust/ntex/Cargo.toml
+++ b/frameworks/Rust/ntex/Cargo.toml
@@ -18,7 +18,7 @@ path = "src/main_raw.rs"
 [dependencies]
 ntex = "0.1.21"
 mimalloc = { version = "0.1.21", default-features = false }
-yarte = { version = "0.12", features = ["bytes-buf", "json"] }
+yarte = { version = "0.12", features = ["json"] }
 env_logger = "0.7"
 random-fast-rng = "0.1.1"
 bytes = "0.5.6"
@@ -33,6 +33,7 @@ serde = { version = "1.0", features = ["derive"] }
 log = { version = "0.4", features = ["release_max_level_off"] }
 tokio = "=0.2.6"
 tokio-postgres = { git="https://github.com/fafhrd91/postgres.git" }
+sailfish = "0.3.0"
 
 [profile.release]
 lto = true