--- a/frameworks/Rust/ntex/templates/fortune.stpl
+++ b/frameworks/Rust/ntex/templates/fortune.stpl
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html>
+  <head><title>Fortunes</title></head>
+  <body>
+    <table>
+      <tr><th>id</th><th>message</th></tr>
+      <% for item in items.iter() { %><tr><td><%= item.id %></td><td><%= &*item.message %></td></tr><% } %>
+    </table>
+  </body>
+</html>