This PR replaces yarte dependency with [sailfish](https://github.com/Kogia-sima/sailfish) crate.

cc https://github.com/Kogia-sima/sailfish/issues/37

## old result

```json
{
  "latencyAvg": "683.48us", 
  "latencyMax": "11.38ms", 
  "latencyStdev": "342.44us", 
  "totalRequests": 356239, 
  "startTime": 1608533180, 
  "endTime": 1608533195
}, 
{
  "latencyAvg": "1.16ms", 
  "latencyMax": "13.34ms", 
  "latencyStdev": "670.61us", 
  "totalRequests": 425133, 
  "startTime": 1608533197, 
  "endTime": 1608533212
}, 
{
  "latencyAvg": "2.01ms", 
  "latencyMax": "17.14ms", 
  "latencyStdev": "1.38ms", 
  "totalRequests": 497662, 
  "startTime": 1608533214, 
  "endTime": 1608533229
}, 
{
  "latencyAvg": "3.63ms", 
  "latencyMax": "27.77ms", 
  "latencyStdev": "2.52ms", 
  "totalRequests": 550767, 
  "startTime": 1608533231, 
  "endTime": 1608533246
}, 
{
  "latencyAvg": "6.49ms", 
  "latencyMax": "36.78ms", 
  "latencyStdev": "3.84ms", 
  "totalRequests": 596106, 
  "startTime": 1608533248, 
  "endTime": 1608533263
}, 
{
  "latencyAvg": "12.05ms", 
  "latencyMax": "56.21ms", 
  "latencyStdev": "5.39ms", 
  "totalRequests": 628707, 
  "startTime": 1608533265, 
  "endTime": 1608533280
}
```

## new result

```json
{
  "latencyAvg": "668.04us", 
  "latencyMax": "6.78ms", 
  "latencyStdev": "312.76us", 
  "totalRequests": 362737, 
  "startTime": 1608534419, 
  "endTime": 1608534434
}, 
{
  "latencyAvg": "1.14ms", 
  "latencyMax": "10.72ms", 
  "latencyStdev": "670.94us", 
  "totalRequests": 431759, 
  "startTime": 1608534436, 
  "endTime": 1608534451
}, 
{
  "latencyAvg": "2.01ms", 
  "latencyMax": "19.30ms", 
  "latencyStdev": "1.42ms", 
  "totalRequests": 503363, 
  "startTime": 1608534453, 
  "endTime": 1608534468
}, 
{
  "latencyAvg": "3.57ms", 
  "latencyMax": "24.58ms", 
  "latencyStdev": "2.49ms", 
  "totalRequests": 560752, 
  "startTime": 1608534470, 
  "endTime": 1608534485
}, 
{
  "latencyAvg": "6.34ms", 
  "latencyMax": "40.29ms", 
  "latencyStdev": "3.65ms", 
  "totalRequests": 608203, 
  "startTime": 1608534487, 
  "endTime": 1608534503
}, 
{
  "latencyAvg": "11.73ms", 
  "latencyMax": "44.56ms", 
  "latencyStdev": "5.17ms", 
  "totalRequests": 644582, 
  "startTime": 1608534505, 
  "endTime": 1608534520
}
```