@botika 2020-12-21
https://github.com/TechEmpower/FrameworkBenchmarks/wiki/Project-Information-Framework-Tests-Overview#fortunes
I think it does not comply with the **XVI**, I think it does not comply with the, since it escapes 4 characters instead of 6 needed in HTML.