@Kogia-sima 2020-12-22
@botika sailfish implementation is based on [ejs](https://github.com/mde/ejs) and escapes **5** characters (&\<\>"'), which is exactly same behavior as ejs.

[Source code](https://github.com/Kogia-sima/sailfish/blob/master/sailfish/src/runtime/escape/mod.rs)

Also I believe many frameworks on this repository escape this **5** characters (such as [lithium](https://github.com/TechEmpower/FrameworkBenchmarks/blob/master/frameworks/C%2B%2B/lithium/lithium.cc), [quicktemplate](https://github.com/valyala/quicktemplate/blob/master/htmlescapewriter.go)).

Also [OWASP Cheat Sheet](https://cheatsheetseries.owasp.org/cheatsheets/Cross_Site_Scripting_Prevention_Cheat_Sheet.html#rule-1-html-encode-before-inserting-untrusted-data-into-html-element-content) states that you should escaped these **5** characters.