@botika 2020-12-22
Good escape algorithm, but `yarte` 0.14 is faster than the previous version, using `Vec` instead of `BytesMut`. 
And I can still do the raw version that you can't imitate with your buffer implementation and relocating 32Kb to 32Kb would go much faster