@nbrady-techempower 2020-12-22
@Kogia-sima The framework maintainer would need to approve a change like this, as he already approved `yarte`. My suggestion would be to create a different permutation, so that `yarte` and `sailfish` can live together and be compared during an actual run. Let me know if you have any questions about how to do that and I can give you some tips. Then it would have to be approved by @fafhrd91